+++ src/main/scala/CesiumApp.scala
@@ -32,18 +32,19 @@ object CesiumApp extends JSApp {
       terrainShadows(ShadowMode.ENABLED))
 
     def createModel(url: String, height: Double) = {
+      val headPitchRol = new HeadingPitchRoll(Math.toRadians(135), 0, 0)
       val position = Cartesian3.fromDegrees(-123.0744619, 44.0503706, height)
       val entity = viewer.entities.add(new Entity(
         EntityOptions.
           name(url).
           position(position).
-          orientation(Transforms.headingPitchRollQuaternion(position, Math.toRadians(135), 0, 0)).
+          orientation(Transforms.headingPitchRollQuaternion(position, headPitchRol)).
           model(new ModelGraphics(ModelGraphicsOptions.
             uri(url).
             minimumPixelSize(128).
             maximumScale(20000))))
       )
-      viewer.trackedEntity = entity
+  //    viewer.trackedEntity = entity
     }
 
     def addBillboard() = {
@@ -106,6 +107,16 @@ object CesiumApp extends JSApp {
     addBillboard()
 
     createModel("./SampleData/models/CesiumAir/Cesium_Air.glb", 5000.0)
+
+    val theImage = new PinBuilder().fromColor(Color.ROYALBLUE, 48).toDataURL("PNG", 1.0)
+
+    viewer.entities.add(new Entity(EntityOptions.
+      name("Blank blue pin").
+      position(Cartesian3.fromDegrees(151.2093, -33.8688)).
+      billboard(new BillboardGraphics(BillboardGraphicsOptions.
+        image(theImage).
+        verticalOrigin(VerticalOrigin.BOTTOM).
+        show(true)))))
   }
 
 }